fix: when user clicks logo it doesn't bring to search TODO: same thing happen when user click MovieCard
